requirements/requirements.txt
# ==============================================================================
# CORE DJANGO DEPENDENCIES
# ==============================================================================
Django==3.2.4
djangorestframework==3.13.1

# ==============================================================================
# AUTHENTICATION & SECURITY
# ==============================================================================
# djangorestframework-simplejwt==4.8.0  # Comentado por conflictos
django-cors-headers==3.7.0
argon2-cffi==21.1.0
PyJWT==2.1.0

# ==============================================================================
# DATABASE
# ==============================================================================
dj-database-url==1.0.0
# psycopg2==2.9.3  # Para PostgreSQL, comentamos por ahora

# ==============================================================================
# API DOCUMENTATION
# ==============================================================================
drf-yasg==1.21.4

# ==============================================================================
# UTILITIES
# ==============================================================================
django-environ==0.4.5
django-filter==22.1
pillow==9.2.0
python-decouple==3.6

# ==============================================================================
# REPORTS & EMAIL (Comentamos por ahora)
# ==============================================================================
# reportlab==3.6.8
# xhtml2pdf==0.2.5
# weasyprint==54.3

# ==============================================================================
# DEPLOYMENT
# ==============================================================================
whitenoise==5.3.0
gunicorn==20.1.0

# ==============================================================================
# DEVELOPMENT & TESTING
# ==============================================================================
# django-debug-toolbar==3.2.4
# django-extensions==3.2.1
# factory-boy==3.2.1
# pytest-django==4.5.2

# # ==============================================================================
# # TASK QUEUE (OPTIONAL)
# # ==============================================================================
# celery==5.2.7
# redis==4.3.4
